requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
requests==2.31.0
scikit-learn==1.4.2
prefect==3.6.26
streamlit==1.35.0
plotly==5.22.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
griffe==1.5.1